+++ cli/js/api_version.js
@@ -1,0 +1,8 @@
+/*
+ * Available Phylum REST API versions.
+ *
+ * NOTE: This must match the type in `extensions/phylum.d.ts`.
+ */
+export const ApiVersion = {
+  V0: "v0",
+};