+++ cli/tests/fixtures/extensions/api/main.ts
@@ -1,4 +1,2 @@
-import { PhylumApi } from "phylum";
-
-const lockfile = await PhylumApi.parseDependencyFile("../tests/fixtures/poetry.lock", "poetry");
+const lockfile = await Phylum.parseDependencyFile("../tests/fixtures/poetry.lock", "poetry");
 console.log(lockfile.packages.length);