+++ extensions/duplicates/main.ts
@@ -2,8 +2,6 @@ import { mapValues } from "https://deno.land/std@0.150.0/collections/map_values.
 import { distinct } from "https://deno.land/std@0.150.0/collections/distinct.ts";
 import { groupBy } from "https://deno.land/std@0.150.0/collections/group_by.ts";
 
-import { PhylumApi } from "phylum";
-
 // Ensure lockfile argument is present.
 if (Deno.args.length != 1) {
   console.error("Usage: phylum duplicates <LOCKFILE>");
@@ -11,7 +9,7 @@ if (Deno.args.length != 1) {
 }
 
 // Parse lockfile using Phylum's API.
-const lockfile = await PhylumApi.parseDependencyFile(Deno.args[0]);
+const lockfile = await Phylum.parseDependencyFile(Deno.args[0]);
 
 // Group all versions for the same dependency together.
 const groupedDeps = groupBy(lockfile.packages, (dep) => dep.name);