+++ extensions/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+- Expose API via global `Phylum` object
+
 ## 6.1.0 - 2024-01-29
 
 - Accept PURLs in `PhylumApi::analyze`